week3/requirements.txt
dotmap==1.3.30
great-expectations==0.15.27
scikit-learn==1.0
hydra-core==1.2.0
wandb==0.13.4
matplotlib==3.2.0
joblib==0.14.1
pathlib==1.0.1
pandas==1.3.4
pytest==6.2.5

